Use correct sequence and metadata when building block
When building a block, we look at the latest round and build the round on top of it,
only if we have already committed a block already.

However, since we notarize rounds asynchronously, we may advance to a new round
before we commit a block.

We therefore need to use the rounds themselves as an indicator.

Signed-off-by: Yacov Manevich <yacov.manevich@avalabs.org>