Prevent re-execution of a block in the same round
This commit prevents re-execution of a block of the same round, twice or more.

This prevents two things:

- A leader sending a block twice to the same node
- A node executing a block twice by receiving one from the leader via the consensus agreement path,
  and also receiving one from the replication code path.

It works by introducing a wrapper around the existing scheduler, which tracks the round number of the block being executed.
This doesn't prevent the same block from being proposed by different nodes in different rounds.

Signed-off-by: Yacov Manevich <yacov.manevich@avalabs.org>