Try to collect an empty notarization after sending an empty vote
A node might be the last node among a quorum that times out.
It therefore receives the empty votes before sending the empty vote itself,
so it should check if it is possible to assemble an empty notarization right after
sending the empty vote, because otherwise it might never get a trigger to check
if an empty notarization can be assembled.

Signed-off-by: Yacov Manevich <yacov.manevich@avalabs.org>