Rebroadcast finalizations for past notarized blocks when finalizing later blocks
Currently, if a node misses a finalization message or a finalization certificate message for a round,
it will not try to re-broadcast the finalization message for that round even if it finalizes a later round.

In our implementation of Simplex, we can only commit a block if we have collected a finalization certificate for the block,
even if a finalization certificate has been collected for a later block.

Also, our implementation of Simplex commits blocks in-order.

It follows from the above that a quorum of correct nodes cannot commit a block if a finalization certificate hasn't been collected for an ancestor block,
because an ancestor block cannot be committed due to missing the finalization certificate.

This commit makes the node re-broadcast finalizations for ancestor blocks of blocks for which a finalization certificate has been collected,
but cannot be committed because it needs to collect finalization certificates for ancestor blocks.

Signed-off-by: Yacov Manevich <yacov.manevich@avalabs.org>