Blacklist implementation
This commit implements a blacklist that performs book-keeping of nodes that are blacklisted.
The blacklist contains a node count, and a bit vector the size of the node count.

When a node is blacklisted, its corresponding bit is set to 1 in the bit vector.

A node can be removed from the blacklist by being voted to be redeemed by f+1 nodes.

The blacklist keeps track of redemptions via a bit vector per blacklisted node.

If a more than f nodes are blacklisted, a random node is removed from the blacklist.

Signed-off-by: Yacov Manevich <yacov.manevich@avalabs.org>