Abort timing out on a stale round
If we have scheduled waiting for pending transactions and then the timeout expired,
in the meantime we may have advanced to a new round.

While a mechanism exists to abort such a situation, it is invoked too early and a race
can occur in which we advance to a new round while we are scheduling the task of the timeout itself.

This commit checks if we have advanced the round right before the timeout triggers, which happens under a lock.

Signed-off-by: Yacov Manevich <yacov.manevich@avalabs.org>