Blacklist implementation (#205)
* Remove redundant code

We currently cannot replicate empty rounds using our message schema,
so this code should be removed.

Signed-off-by: Yacov Manevich <yacov.manevich@avalabs.org>

* Blacklist implementation for skipping crashed leaders

This commit implements a blacklist mechanism.
Nodes that crash are blacklisted if rounds in which it is their turn to propose a block do not end up with a block.
Nodes that are blacklisted cannot become leaders until they redeem themselves by notarizing blocks.

Blacklisting or redeeming a node requires f+1 votes from nodes. Nodes vote by proposing blacklist updates in blocks they propose.
A blacklist cannot extend beyond f nodes. Nodes need to be redeemed and removed from the blacklist before additional nodes can be added to it.

When a node is inserted into the blacklist, its orbit is recorded. In order for the node to be blacklisted, f+1 or more nodes need to accuse it
within n rounds, otherwise the accusations are garbage collected. This is to prevent from a rogue node to repeatedly voting to blacklist a node
across many rounds. Similarly, in order for a node to be redeemed and removed from the blacklist, f+1 or more nodes need to vote to redeem it
within n rounds.

Signed-off-by: Yacov Manevich <yacov.manevich@avalabs.org>

* Some last nits

Signed-off-by: Yacov Manevich <yacov.manevich@avalabs.org>

---------

Signed-off-by: Yacov Manevich <yacov.manevich@avalabs.org>