Mark block as being processed, not as a future block
When we receive a block, we schedule it to be processed:
- Verify the vote
- Execute/Verify the block
- Write to to the WAL
etc.

We may receive a finalization in the meanwhile, and we therefore store it as a future message to be processed.

Since we received a finalization for a future round, this might trigger the replication process.
In order to prevent us from initiating replication needlessly, Simplex would store the pending block message as a future message,
and then when a finalization is received, it would look if a block message exists among the future messages, as an indication for the block
being processed.

This works, but creates an annoying downside - the block message is still in the future messages, and it may get picked up once more
by anything that optimistically calls maybeLoadFutureMessages(), and will needlessdly try to re-verify the block.

This commit replaces the mechanism with an explicit boolean instead of placing the entire block message in the future messages,
which prevents the block message from being handled more than once per its reception.

Signed-off-by: Yacov Manevich <yacov.manevich@avalabs.org>