Kill zombie tasks when finalizing
This commit adds a functionality to the task scheduler that enables to kill tasks not yet ready to run,
and kills these tasks if they will never be executed, in order to avoid these tasks from residing in the memory for ever.

Since blocks are verified in-order, it is only possible to verify a block if its parent block has been verified.
Therefore, a task that verifies a block can be not yet ready to run only if its parent block has not yet been verified.

This commit makes the Simplex instance kill all tasks that correspond to sequences that have been already finalized,
as a block that corresponds to a sequence that is finalized, does not need to be verified.

Signed-off-by: Yacov Manevich <yacov.manevich@avalabs.org>